Creates a new folder called 'router' and nests a new file inside of it called 'router.js'
